cleanup app.js
