+++ src/screens/HomeScreen.js
@@ -1,0 +1,66 @@
+import React from 'react';
+import {
+  SafeAreaView,
+  StyleSheet,
+  ScrollView,
+  View,
+  Text,
+  StatusBar,
+} from 'react-native';
+import {
+    Header,
+    LearnMoreLinks,
+    Colors,
+    DebugInstructions,
+    ReloadInstructions,
+  } from 'react-native/Libraries/NewAppScreen';
+
+const HomeScreen = () => {
+    return(
+        <SafeAreaView>
+            <Text>Test is the test tests </Text>
+        </SafeAreaView>
+    )
+}
+
+const styles = StyleSheet.create({
+    scrollView: {
+      backgroundColor: Colors.lighter,
+    },
+    engine: {
+      position: 'absolute',
+      right: 0,
+    },
+    body: {
+      backgroundColor: Colors.white,
+    },
+    sectionContainer: {
+      marginTop: 32,
+      paddingHorizontal: 24,
+    },
+    sectionTitle: {
+      fontSize: 24,
+      fontWeight: '600',
+      color: Colors.black,
+    },
+    sectionDescription: {
+      marginTop: 8,
+      fontSize: 18,
+      fontWeight: '400',
+      color: Colors.dark,
+    },
+    highlight: {
+      fontWeight: '700',
+    },
+    footer: {
+      color: Colors.dark,
+      fontSize: 12,
+      fontWeight: '600',
+      padding: 4,
+      paddingRight: 12,
+      textAlign: 'right',
+    },
+  });
+  
+
+export default HomeScreen
